utils/cli/requirements.txt
Pillow>=6.2.0
requests>=2.20.1
google-cloud-storage>=1.28.1